+++ src/components/user-action/input/InsideLabelInput.tsx
@@ -47,8 +47,8 @@ export const InsideLabelInput = forwardRef<HTMLInputElement, InsideLabelInputPro
         data-display={isFocused || !!value ? 'small' : 'full'}
         className={clsx(
           'absolute left-4 top-2 transition-all pointer-events-none touch-none',
-          'data-[display=small]:top-2 data-[display=small]:h-force-4.5 data-[display=small]:typography-caption-sm-regular data-[display=small]:overflow-y-hidden',
-          'data-[display=full]:top-1/2 data-[display=full]:-translate-y-1/2 data-[display=full]:typography-body-md-regular'
+          'data-[display=small]:top-2 data-[display=small]:h-force-4.5 data-[display=small]:typography-caption-sm data-[display=small]:overflow-y-hidden',
+          'data-[display=full]:top-1/2 data-[display=full]:-translate-y-1/2 data-[display=full]:typography-body-md'
         )}
       >
         {label}